+++ backend/config/settings.py
@@ -40,6 +40,7 @@ INSTALLED_APPS = [
     'apps.datasource',  # 数据源管理模块
     'apps.dataasset',  # 数据资产管理模块（包含元数据管理）
     'apps.dataetl',     # ETL任务管理模块
+    'apps.datadev',     # 数据开发模块
     'apps.terminal',    # Web Terminal
 ]